Add a feature to let me turn the display on and off over the wire
Also increase the error checking on the config, since it might not
always be accurate in MQTT.

This change has been worked into my scenes, too. Now when it's time to
wind down the display gets dim and the colon stops blinking, and when
it's movie time the display turns off completely!